+++ wishlist-service/src/main/java/com/microservices/customer/entity/Customer.java
@@ -1,6 +1,7 @@
 package com.microservices.customer.entity;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
@@ -9,6 +10,8 @@ import lombok.Setter;
 @Setter
 @Entity
 @Table
-public class Customer extends MyUser {
-
+public class Customer {
+	@Id
+	private long id;
+	private String username;
 }
