+++ wishlist-service/src/main/java/com/microservices/customer/entity/Product.java
@@ -2,8 +2,6 @@ package com.microservices.customer.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Getter;
@@ -15,7 +13,6 @@ import lombok.Setter;
 @Setter
 public class Product {
 	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private long id;
 	@Column(nullable = false)
 	private String title;
