+++ wishlist-service/src/main/resources/application.properties
@@ -15,3 +15,5 @@ logging.level.org.springframework.cloud.loadbalancer=ERROR
 # You can also suppress Spring's BeanPostProcessorChecker warnings
 logging.level.org.springframework.beans.factory.support=ERROR
 spring.datasource.hikari.connection-timeout=60000
+
+spring.h2.console.enabled=true
